+++ example/shared/Definitions.ts
@@ -1,19 +1,18 @@
-import { NetBuilder, DefinitionBuilder, RateLimiter, Serialization } from "@rbxts/netbuilder";
+import { NetBuilder, DefinitionBuilder, RateLimiter } from "@rbxts/netbuilder";
 import { t } from "@rbxts/t";
 
 import Person from "./Class/Person";
 import Log from "./Util/Log";
 
-export = new NetBuilder()
-	.Configure((config) =>
-		config
-			.SetRootName("ExampleRemotes")
-			.SetLogger({
-				Error: (def, stderr) => Log.Error(`[${def.Id}] ${stderr}`),
-				Warn: (def, ...params) => Log.Warn(`[${def.Id}]`, ...params),
-			})
-			.CacheFunctions(),
-	)
+const remotes = new NetBuilder()
+	.Configure({
+		RootName: "ExampleRemotes",
+		Logger: {
+			Error: (def, stderr) => Log.Error(`[${def.Id}] ${stderr}`),
+			Warn: (def, ...params) => Log.Warn(`[${def.Id}]`, ...params),
+		},
+		CacheFunctions: true,
+	})
 	.BindDefinition(new DefinitionBuilder("PrintOnClient").SetArguments(t.string).Build())
 	.BindNamespace(
 		"People",
@@ -29,3 +28,7 @@ export = new NetBuilder()
 			.AsNamespace(),
 	)
 	.Build();
+
+export const Server = remotes.Server;
+
+export const Client = remotes.Client;