+++ src/Util/createLoggingDefinition.ts
@@ -1,15 +1,11 @@
 import { Definition, DefinitionMembers, LoggingDefinition } from "../definitions";
 
 /** @internal */
-function createLoggingDefinition(definition: Definition | DefinitionMembers) {
-	const loggingDefinition: LoggingDefinition = {
+function createLoggingDefinition(definition: Definition | DefinitionMembers): LoggingDefinition {
+	return table.freeze({
 		Id: (definition as DefinitionMembers).Id,
 		Kind: (definition as DefinitionMembers).Kind,
-	};
-
-	table.freeze(loggingDefinition);
-
-	return loggingDefinition;
+	});
 }
 
 export = createLoggingDefinition;