+++ src/Util/definitionInfo.ts
@@ -1,7 +1,7 @@
-import { Definition, DefinitionMembers } from "../definitions";
+import { Definition, DefinitionMembers, LoggingDefinition } from "../definitions";
 import { IS_SERVER } from "./constants";
 
-export = (definition: Definition | DefinitionMembers, includeBoundary = false) =>
+export = (definition: Definition | LoggingDefinition, includeBoundary = false) =>
 	`${includeBoundary ? (IS_SERVER ? "Server" : "Client") : ""}${
 		(definition as DefinitionMembers).Kind
 	}<${(definition as DefinitionMembers).Id}>`;