+++ CHANGELOG.md
@@ -4,13 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [0.5.0] - 2022-05-20
+## [0.6.0] - 2022-05-30
+### Added
+- Added a `ConnectParallel` method to both dispatchers.
+- Exposed the `Check` utility type.
+
+### Changed
+- Internal changes
+- Improved the generated remotes types for better visualization.
+- More JSDoc texts to some object fields for readability.
+
+### Fixed
+- Fixed object type configuration in `NetBuilder:Configure` not working applying the settings.
+
+## [0.5.0] - 2022-05-20
 ### Added
 - Added `PreGeneration` configuration field.
 - Added `CacheFunctions` configuration field.
 - Added `Label` configuration field.
 - Added `Debug` configuration field.
-- Added a `Wait` method for both dispatchers.
+- Added a `Wait` method to both dispatchers.
 
 ### Changed
 - **[breaking]** Changed `SupressWarnings` method in `ConfigurationBuilder` to `SuppressWarnings`. (typo fix)
@@ -22,13 +35,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Removed deprecated configuration methods from `NetBuilder`.
 
-
-# [0.4.0] - 2022-04-30
+## [0.4.0] - 2022-04-30
 ### Added
 - Added `RootName` field to configuration.
 - Added `SetRootName` method to `ConfigurationBuilder`.
 
-# [0.3.0] - 2022-04-29
+## [0.3.0] - 2022-04-29
 ### Added
 - Added `Serialization.CreateSerializationDefinition`.
 - Added `SerializationDefinition`.